// src/modules/general/Ordenes/types/cantidadEtapasOrden.types.ts
export interface ICantidadEtapasOrden {
    cantidad_giradas: number
    cantidad_pre_presupuestadas: number
    cantidad_presupuestadas: number
    cantidad_aceptadas: number
    cantidad_dinero_entregado: number
    cantidad_lista_realizar: number
    cantidad_descargadas: number
    cantidad_pronuncio_abogado: number
    cantidad_cuentas_conciliadas: number
    cantidad_vencidas_leves: number
    cantidad_vencidas_graves: number
  }